requirements.txt
scikit-surprise
scikit-learn
numpy
pandas
